+++ Plugins/Creature/Creature.hpp
@@ -138,6 +138,8 @@ private:
     ArgumentStack SetPersonalSpace              (ArgumentStack&& args);
     ArgumentStack GetCreaturePersonalSpace      (ArgumentStack&& args);
     ArgumentStack SetCreaturePersonalSpace      (ArgumentStack&& args);
+    ArgumentStack GetHeight                     (ArgumentStack&& args);
+    ArgumentStack SetHeight                     (ArgumentStack&& args);
 
     CNWSCreature *creature(ArgumentStack& args);
     std::unordered_map<uint8_t, std::unordered_map<ObjectID, int16_t>> m_RollModifier;